+++ dbbuild/sources/puentephillips2020/compile_marker_definitions.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+#
 # -----------------------------------------------------------------------------
 # Copyright (c) 2020, Battelle National Biodefense Institute.
 #
@@ -5,10 +7,10 @@
 # and is licensed under the BSD license: see LICENSE.txt.
 # -----------------------------------------------------------------------------
 
+import argparse
 import builtins
-from collections import defaultdict
 from gzip import open as gzopen
-import sys
+import pandas
 
 
 def smartopen(filename, mode):
@@ -29,7 +31,7 @@ def smartopen(filename, mode):
     return openfunc(filename, mode)
 
 
-def marker_coords(vcf, mapping):
+def parse_rsid_coords(vcf):
     rsidcoords = dict()
     for line in vcf:
         if line.startswith('#'):
@@ -38,18 +40,53 @@ def marker_coords(vcf, mapping):
         chrom = chrnum if chrnum.startswith('chr') else 'chr' + chrnum
         pos = int(posstr) - 1
         rsidcoords[rsid] = (chrom, pos)
+    return rsidcoords
 
-    marker_rsids = dict()
-    next(mapping)
-    for line in mapping:
-        marker, pos, span, rsidlist, avggd = line.strip().split('\t')
-        marker_rsids[marker] = rsidlist.split(',')
 
-    for marker, rsids in marker_rsids.items():
+def marker_coords(markerfile, rsidcoords):
+    final_defs = {
+        'Name': list(),
+        'Xref': list(),
+        'NumVars': list(),
+        'Chrom': list(),
+        'OffsetsHg37': list(),
+        'OffsetsHg38': list(),
+        'VarRef': list(),
+    }
+    markers = pandas.read_csv(markerfile, sep='\t')
+    for n, row in markers.iterrows():
+        rsids = row.RSIDs.split(',')
         positions = [rsidcoords[r][1] for r in rsids]
         rsids = [rsid for pos, rsid in sorted(zip(positions, rsids))]
         positions = sorted(positions)
+        offsetstr = ','.join(map(str, positions))
         chrom = rsidcoords[rsids[0]][0]
         chromlabel = '0X' if chrom == 'chrX' else '{:02d}'.format(int(chrom[3:]))
-        mhname = 'mh{chr:s}USC-{n:s}'.format(chr=chromlabel, n=marker)
-        yield mhname, chrom, positions, rsids
+        name = f'mh{chromlabel}USC-{row.Label}'
+        final_defs['Name'].append(name)
+        final_defs['Xref'].append(None)
+        final_defs['NumVars'].append(len(positions))
+        final_defs['Chrom'].append(chrom)
+        final_defs['OffsetsHg37'].append(None)
+        final_defs['OffsetsHg38'].append(offsetstr)
+        final_defs['VarRef'].append(','.join(rsids))
+    return pandas.DataFrame(final_defs)
+
+
+def cli():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('markerrsids')
+    parser.add_argument('vcf')
+    parser.add_argument('out')
+    return parser
+
+
+def main(args):
+    with smartopen(args.vcf, 'r') as fh:
+        rsidcoords = parse_rsid_coords(fh)
+    markers = marker_coords(args.markerrsids, rsidcoords)
+    markers.to_csv(args.out, sep='\t', index=False)
+
+
+if __name__ == '__main__':
+    main(cli().parse_args())