+++ dbbuild/sources/voskoboinik2018/compile_marker_definitions.py
@@ -1,6 +1,13 @@
 #!/usr/bin/env python
+#
+# -----------------------------------------------------------------------------
+# Copyright (c) 2020, Battelle National Biodefense Institute.
+#
+# This file is part of MicroHapDB (http://github.com/bioforensics/microhapdb)
+# and is licensed under the BSD license: see LICENSE.txt.
+# -----------------------------------------------------------------------------
+
 import argparse
-import json
 import os
 import pandas
 import re
@@ -57,17 +64,13 @@ def parse_vcf(infile):
         yield Variant(line)
 
 
-def get_regions(table1file):
+def parse_marker_definitions(table1file):
     data = pandas.read_csv(table1file, sep='\t')
     data['End'] = data['GRCh37Pos'] + data['Length']
     for n, row in data.iterrows():
-        yield '{chr:d}:{start:d}-{end:d}'.format(chr=row.Chrom, start=row.GRCh37Pos, end=row.End)
-
-
-def get_marker_names(table1file):
-    data = pandas.read_csv(table1file, sep='\t')
-    for n, row in data.iterrows():
-        yield 'mh{chr:02d}LV-{sn:02d}'.format(chr=row.Chrom, sn=n+1)
+        name = f'mh{row.Chrom:02d}LV-{n+1:02d}'
+        region = f'{row.Chrom}:{row.GRCh37Pos}-{row.End}'
+        yield name, region
 
 
 def get_variants(markername, region, path):
@@ -106,7 +109,7 @@ def filter_variants(markername):
             print(var, file=outfh)
 
 
-def transfer_variants(markername, vcf, rsidx):
+def transfer_variants(markername, vcf, rsidx, refr='GRCh38'):
     infile = '{mn:s}-filtered.vcf'.format(mn=markername)
     rsids = list()
     with open(infile, 'r') as fh:
@@ -116,61 +119,72 @@ def transfer_variants(markername, vcf, rsidx):
             values = line.split('\t')
             rsid = values[2]
             rsids.append(rsid)
-    outfile = '{mn:s}-GRCh38.vcf'.format(mn=markername)
+    outfile = f'{markername}-{refr}.vcf'
     command = ['rsidx', 'search', vcf, rsidx] + rsids
     with open(outfile, 'w') as fh:
         subprocess.check_call(command, universal_newlines=True, stdout=fh)
+
+
+def vcf_to_rsid_offsets(vcfstream):
+    offsets = list()
+    rsids = list()
+    for line in vcfstream:
+        chrom, pos, rsid, *values = line.strip().split('\t')
+        markerchrom = 'chr' + chrom
+        offsets.append(int(pos) - 1)
+        rsids.append(rsid)
+    return markerchrom, rsids, offsets
 
 
 def finalize_marker_definitions(markernames):
     data = {
         'Name': list(),
         'Xref': list(),
-        'Reference': list(),
+        'NumVars': list(),
         'Chrom': list(),
-        'Offsets': list(),
+        'OffsetsHg37': list(),
+        'OffsetsHg38': list(),
         'VarRef': list(),
     }
     for name in markernames:
         infile = '{mn:s}-GRCh38.vcf'.format(mn=name)
         offsets = list()
         rsids = list()
-        with open(infile, 'r') as fh:
-            for line in fh:
-                chrom, pos, rsid, *values = line.strip().split('\t')
-                markerchrom = 'chr' + chrom
-                offsets.append(int(pos) - 1)
-                rsids.append(rsid)
-        offsetstr = ','.join(map(str, offsets))
-        rsidstr = ','.join(rsids)
+        with open(f'{name}-GRCh37.vcf', 'r') as fh:
+            chrom, rsids37, offsets37 = vcf_to_rsid_offsets(fh)
+        with open(f'{name}-GRCh38.vcf', 'r') as fh:
+            chrom, rsids38, offsets38 = vcf_to_rsid_offsets(fh)
+        assert sorted(rsids37) == sorted(rsids38), (sorted(rsids37), sorted(rsids38))
         data['Name'].append(name)
         data['Xref'].append(None)
-        data['Reference'].append('GRCh38')
-        data['Chrom'].append(markerchrom)
-        data['Offsets'].append(offsetstr)
-        data['VarRef'].append(rsidstr)
+        data['NumVars'].append(len(rsids38))
+        data['Chrom'].append(chrom)
+        data['OffsetsHg37'].append(','.join(map(str, offsets37)))
+        data['OffsetsHg38'].append(','.join(map(str, offsets38)))
+        data['VarRef'].append(','.join(rsids38))
     df = pandas.DataFrame(data)
     df.to_csv('marker.tsv', sep='\t', index=False)
 
 
 def cli():
     parser = argparse.ArgumentParser()
-    parser.add_argument('--configfile', default='../../config.json')
-    parser.add_argument('data', help='Table 1 from Voskoboinik et al 2018')
+    parser.add_argument('dbsnp37', help='dbSNP VCF for GRCh37/hg19')
+    parser.add_argument('dbsnp38', help='dbSNP VCF for GRCh38')
+    parser.add_argument('rsidx37', help='rsidx index for GRCh37/hg19')
+    parser.add_argument('rsidx38', help='rsidx index for GRCh38')
+    parser.add_argument('dir1kgp', help='directory containing 1000 Genomes Project VCFs')
+    parser.add_argument('table1', help='Table 1 from Voskoboinik et al 2018')
     return parser
 
 
 def main(args):
-    with open(args.configfile, 'r') as fh:
-        config = json.load(fh)
-        vcf = config['dbsnp38']
-        rsidx = config['dbsnp38'].replace('.vcf.gz', '.rsidx')
-    markernames = list(get_marker_names(args.data))
-    regions = list(get_regions(args.data))
-    for name, region in zip(markernames, regions):
-        get_variants(name, region, config['1kgp_dir'])
+    markernames = list()
+    for name, region in parse_marker_definitions(args.table1):
+        markernames.append(name)
+        get_variants(name, region, args.dir1kgp)
         filter_variants(name)
-        transfer_variants(name, vcf, rsidx)
+        transfer_variants(name, args.dbsnp37, args.rsidx37, refr='GRCh37')
+        transfer_variants(name, args.dbsnp38, args.rsidx38, refr='GRCh38')
     finalize_marker_definitions(markernames)